requirements.txt
black==22.12.0
mypy==0.991
pandas==1.5.2
pre-commit==2.21.0
PyGithub==1.57
pylint==2.15.9
pytest==7.2.0
tabulate==0.9.0
